backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
google-genai==0.3.0
python-multipart==0.0.9
PyPDF2==3.0.1
python-dotenv==1.0.1
pydantic==2.6.1
aiofiles==23.2.1